Separated out the controllers into Dialogue/Movement/Jump
