+++ Assets/_Game/Scenes/Test Scenes/Monys Test Scene.unity
@@ -767,6 +767,8 @@ GameObject:
   - component: {fileID: 990268614}
   - component: {fileID: 990268613}
   - component: {fileID: 990268612}
+  - component: {fileID: 990268618}
+  - component: {fileID: 990268617}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -839,15 +841,6 @@ MonoBehaviour:
   isInteracting: 0
   isMovementDisabled: 0
   dialogueManager: {fileID: 0}
-  moveSpeed: 3
-  jumpsAllowed: 2
-  jumpForce: 8
-  isGrounded: 1
-  whatIsGround:
-    serializedVersion: 2
-    m_Bits: 64
-  checkRadius: 0.6
-  movement: {x: 0, y: 0}
 --- !u!212 &990268615
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -915,6 +908,39 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &990268617
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 990268611}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f5d804e0c67f44db98ff2c5ceace2e15, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  moveSpeed: 3
+  movement: {x: 0, y: 0}
+--- !u!114 &990268618
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 990268611}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3bfbec78a63a4060b3a498efa3af5546, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  jumpsAllowed: 2
+  jumpForce: 8
+  isGrounded: 1
+  whatIsGround:
+    serializedVersion: 2
+    m_Bits: 64
+  checkRadius: 0.6
 --- !u!1001 &1434375408
 PrefabInstance:
   m_ObjectHideFlags: 0
